pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "isort>=6.1.0",
    "langchain-core>=0.3.83",
    "langchain-openai>=0.3.35",
    "langsmith>=0.4.37",
    "openai>=2.29.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.23",
]
